refactor: extract site_explorer and firmware into dedicated crates
Carve site_explorer out of the carbide-api mega-crate into a new
carbide-site-explorer crate. Along the way, extract the FirmwareConfig
loader into a new carbide-firmware crate — site_explorer depends on
it, and leaving it inside carbide-api would have pulled carbide-api
back in as a transitive dep.

New crates:

- carbide-firmware: FirmwareConfig loader + merge logic. Data types
  (Firmware, FirmwareEntry, FirmwareComponentType, ...) stay in
  carbide-api-model where they were already consumed. Minimal deps:
  bmc-vendor, api-model, eyre, serde, toml, tracing.

- carbide-site-explorer: the full site_explorer module, including
  bmc_endpoint_explorer, machine_creator, switch_creator, config,
  errors, metrics, managed_host helpers, and test fixtures. Depends
  on redfish, ipmi, network, ssh, utils, firmware, secrets, and the
  api-model / api-db data layers — zero carbide-api coupling.

carbide-api now depends on both new crates. All call sites updated:
cfg/file.rs (FirmwareConfig construction), handlers, preingestion,
state_controller, machine_update_manager, setup, tests.

FirmwareConfig gains a `new(firmware_dir, host_models, dpu_models)`
constructor so CarbideConfig::get_firmware_config() shrinks to a
single call. The nv-redfish workspace dep and the `recursion_limit = 256`
crate attribute move from carbide-api to carbide-site-explorer, where
they are actually used.

bmc-explorer-cli now depends on carbide-site-explorer directly instead
of going through carbide-api.

Part of the effort to break up the carbide-api mega-crate. Combined
with the earlier redfish, utils, metrics, config, and errors
extractions, site_explorer now lives on its own and is ready to be
consumed by other binaries without pulling in the full carbide-api
compile graph.
