+++ crates/api/src/machine_update_manager/host_firmware.rs
@@ -21,6 +21,7 @@ use std::sync::Arc;
 use std::sync::atomic::{AtomicU64, Ordering};
 
 use async_trait::async_trait;
+use carbide_firmware::FirmwareConfig;
 use carbide_uuid::machine::MachineId;
 use db::{self, desired_firmware};
 use model::machine::ManagedHostStateSnapshot;
@@ -31,7 +32,7 @@ use tokio::sync::Mutex;
 
 use super::machine_update_module::MachineUpdateModule;
 use crate::CarbideResult;
-use crate::cfg::file::{CarbideConfig, FirmwareConfig};
+use crate::cfg::file::CarbideConfig;
 
 pub struct HostFirmwareUpdate {
     pub metrics: HostFirmwareUpdateMetrics,
