+++ crates/api/src/preingestion_manager/mod.rs
@@ -21,7 +21,9 @@ use std::io;
 use std::sync::Arc;
 use std::time::Duration;
 
+use carbide_firmware::FirmwareConfig;
 use carbide_redfish::libredfish::{RedfishClientCreationError, RedfishClientPool};
+use carbide_site_explorer::EndpointExplorer;
 use chrono::{DateTime, Utc};
 use db::work_lock_manager::WorkLockManagerHandle;
 use db::{DatabaseError, WithTransaction};
@@ -43,10 +45,9 @@ use tokio::task::JoinSet;
 use tokio_util::sync::CancellationToken;
 use utils::periodic_timer::PeriodicTimer;
 
-use crate::cfg::file::{CarbideConfig, FirmwareConfig, FirmwareGlobal};
+use crate::cfg::file::{CarbideConfig, FirmwareGlobal};
 use crate::firmware_downloader::FirmwareDownloader;
 use crate::preingestion_manager::metrics::PreingestionMetrics;
-use crate::site_explorer::EndpointExplorer;
 use crate::{CarbideError, CarbideResult};
 
 mod metrics;
