+++ crates/api/src/tests/machine_creator.rs
@@ -20,6 +20,9 @@ use std::net::IpAddr;
 use std::str::FromStr;
 use std::sync::Arc;
 
+use carbide_site_explorer::MachineCreator;
+use carbide_site_explorer::config::SiteExplorerConfig;
+use carbide_site_explorer::errors::SiteExplorerError;
 use carbide_uuid::machine::MachineId;
 use itertools::Itertools;
 use mac_address::MacAddress;
@@ -35,9 +38,6 @@ use tonic::Request;
 use utils::models::arch::CpuArchitecture;
 
 use crate::cfg::file::DpuConfig as InitialDpuConfig;
-use crate::site_explorer::MachineCreator;
-use crate::site_explorer::config::SiteExplorerConfig;
-use crate::site_explorer::errors::SiteExplorerError;
 use crate::state_controller::machine::handler::MachineStateHandlerBuilder;
 use crate::tests::common;
 use crate::tests::common::api_fixtures::TestEnvOverrides;
