+++ crates/api/src/tests/mod.rs
@@ -19,7 +19,6 @@ mod compute_allocation;
 mod connected_device;
 mod create_domain;
 mod credential;
-mod desired_firmware_versions;
 mod dhcp_lease_expiration;
 mod dns;
 mod dpa_interfaces;
