+++ crates/bmc-explorer-cli/Cargo.toml
@@ -8,7 +8,7 @@ authors.workspace = true
 [dependencies]
 # [local-dependencies]
 # DO NOT PUT DEPENDENCIES OTHER THAN LOCAL DEPS HERE, THEY SHOULD ALL HAVE 'path =' IN THEM.
-carbide-api = { path = "../api", default-features = false }
+carbide-site-explorer = { path = "../site-explorer" }
 carbide-secrets = { path = "../secrets" }
 carbide-ipmi = { path = "../ipmi" }
 carbide-redfish = { path = "../redfish" }
