+++ crates/bmc-explorer-cli/src/main.rs
@@ -18,8 +18,9 @@ use std::sync::Arc;
 use std::time::Instant;
 
 use arc_swap::ArcSwap;
-use carbide::{BmcEndpointExplorer, SiteExplorerExploreMode};
 use carbide_redfish::nv_redfish::NvRedfishClientPool;
+use carbide_site_explorer::BmcEndpointExplorer;
+use carbide_site_explorer::config::SiteExplorerExploreMode;
 use clap::Parser;
 use forge_secrets::credentials::{Credentials, TestCredentialManager};
 use mac_address::MacAddress;
