+++ crates/firmware/src/tests/desired_firmware_versions.rs
@@ -16,14 +16,14 @@
  */
 use sqlx::FromRow;
 
-use crate::cfg::file::FirmwareConfig;
+use crate::FirmwareConfig;
 
 #[derive(FromRow)]
 struct AsStrings {
     pub versions: String,
 }
 
-#[crate::sqlx_test]
+#[super::sqlx_test]
 pub async fn test_build_versions(pool: sqlx::PgPool) -> Result<(), eyre::Error> {
     use std::ops::DerefMut;
 
@@ -116,6 +116,7 @@ current_version_reported_as = "^2$"
     let query = r#"SELECT explicit_update_start_needed FROM desired_firmware;"#;
     let explicit_update_needed: (bool,) = sqlx::query_as(query).fetch_one(txn.deref_mut()).await?;
     assert_eq!(explicit_update_needed, (true,));
+    txn.commit().await?;
 
     Ok(())
 }
