+++ crates/site-explorer/src/errors.rs
