+++ crates/site-explorer/src/test_data/bf2_report.json
