+++ crates/site-explorer/src/test_data/dell_report.json
